handle: add global status handler (#46974)
ref pingcap/tidb#46905
