+++ server/handler/optimizor/BUILD.bazel
@@ -56,6 +56,7 @@ go_test(
         "//server/internal/util",
         "//session",
         "//statistics/handle",
+        "//statistics/handle/globalstats",
         "//store/mockstore/unistore",
         "//testkit",
         "//testkit/testsetup",
