+++ statistics/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
         "histogram.go",
         "index.go",
         "interact_with_storage.go",
-        "merge_worker.go",
         "row_sampler.go",
         "sample.go",
         "scalar.go",
@@ -63,7 +62,6 @@ go_test(
     name = "statistics_test",
     timeout = "short",
     srcs = [
-        "cmsketch_bench_test.go",
         "cmsketch_test.go",
         "fmsketch_test.go",
         "histogram_bench_test.go",
@@ -100,7 +98,6 @@ go_test(
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
-        "@com_github_tiancaiamao_gp//:gp",
         "@org_uber_go_goleak//:goleak",
     ],
 )
