+++ statistics/handle/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//sessiontxn",
         "//statistics",
         "//statistics/handle/cache",
+        "//statistics/handle/globalstats",
         "//statistics/handle/lockstats",
         "//statistics/handle/metrics",
         "//table",
@@ -76,6 +77,7 @@ go_test(
         "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//statistics",
+        "//statistics/handle/globalstats",
         "//statistics/handle/internal",
         "//testkit",
         "//testkit/testsetup",
