+++ statistics/handle/handle.go
@@ -39,6 +39,7 @@ import (
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/statistics"
 	"github.com/pingcap/tidb/statistics/handle/cache"
+	"github.com/pingcap/tidb/statistics/handle/globalstats"
 	handle_metrics "github.com/pingcap/tidb/statistics/handle/metrics"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/types"
@@ -53,18 +54,8 @@ import (
 	"go.uber.org/zap"
 )
 
-const (
-	// TiDBGlobalStats represents the global-stats for a partitioned table.
-	TiDBGlobalStats = "global"
-	// MaxPartitionMergeBatchSize indicates the max batch size for a worker to merge partition stats
-	MaxPartitionMergeBatchSize = 256
-)
-
 // Handle can update stats info periodically.
 type Handle struct {
-	// this gpool is used to reuse goroutine in the mergeGlobalStatsTopN.
-	gpool *gp.Pool
-
 	pool sessionPool
 
 	// initStatsCtx is the ctx only used for initStats
@@ -97,6 +88,8 @@ type Handle struct {
 
 	// statsUsage contains all the column stats usage information from collectors when we dump them to KV.
 	statsUsage *statsUsage
+
+	globalstatushandler *globalstats.GlobalStatusHandler
 
 	// StatsLoad is used to load stats concurrently
 	StatsLoad StatsLoad
@@ -192,8 +185,9 @@ type sessionPool interface {
 // NewHandle creates a Handle for update stats.
 func NewHandle(ctx, initStatsCtx sessionctx.Context, lease time.Duration, pool sessionPool, tracker sessionctx.SysProcTracker, autoAnalyzeProcIDGetter func() uint64) (*Handle, error) {
 	cfg := config.GetGlobalConfig()
+	gpool := gp.New(math.MaxInt16, time.Minute)
 	handle := &Handle{
-		gpool:                   gp.New(math.MaxInt16, time.Minute),
+		globalstatushandler:     globalstats.NewGlobalStatusHandler(gpool),
 		ddlEventCh:              make(chan *ddlUtil.Event, 1000),
 		listHead:                NewSessionStatsCollector(),
 		idxUsageListHead:        &SessionIndexUsageCollector{mapper: make(indexUsageMap)},
@@ -341,52 +335,12 @@ func (h *Handle) UpdateSessionVar() error {
 	return err
 }
 
-// GlobalStats is used to store the statistics contained in the global-level stats
-// which is generated by the merge of partition-level stats.
-// It will both store the column stats and index stats.
-// In the column statistics, the variable `num` is equal to the number of columns in the partition table.
-// In the index statistics, the variable `num` is always equal to one.
-type GlobalStats struct {
-	Hg                    []*statistics.Histogram
-	Cms                   []*statistics.CMSketch
-	TopN                  []*statistics.TopN
-	Fms                   []*statistics.FMSketch
-	MissingPartitionStats []string
-	Num                   int
-	Count                 int64
-	ModifyCount           int64
-}
-
 // MergePartitionStats2GlobalStatsByTableID merge the partition-level stats to global-level stats based on the tableID.
 func (h *Handle) MergePartitionStats2GlobalStatsByTableID(sc sessionctx.Context,
 	opts map[ast.AnalyzeOptionType]uint64, is infoschema.InfoSchema,
 	physicalID int64, isIndex int, histIDs []int64,
-	tablePartitionStats map[int64]*statistics.Table) (globalStats *GlobalStats, err error) {
-	// get the partition table IDs
-	globalTable, ok := h.getTableByPhysicalID(is, physicalID)
-	if !ok {
-		err = errors.Errorf("unknown physical ID %d in stats meta table, maybe it has been dropped", physicalID)
-		return
-	}
-	globalTableInfo := globalTable.Meta()
-	globalStats, err = h.mergePartitionStats2GlobalStats(sc, opts, is, globalTableInfo, isIndex, histIDs, tablePartitionStats)
-	if err != nil {
-		return
-	}
-	if len(globalStats.MissingPartitionStats) > 0 {
-		var item string
-		if isIndex == 0 {
-			item = "columns"
-		} else {
-			item = "index"
-			if len(histIDs) > 0 {
-				item += " " + globalTableInfo.FindIndexNameByID(histIDs[0])
-			}
-		}
-		logutil.BgLogger().Warn("missing partition stats when merging global stats", zap.String("table", globalTableInfo.Name.L),
-			zap.String("item", item), zap.Strings("missing", globalStats.MissingPartitionStats))
-	}
-	return
+	tablePartitionStats map[int64]*statistics.Table) (globalStats *globalstats.GlobalStats, err error) {
+	return h.globalstatushandler.MergePartitionStats2GlobalStatsByTableID(sc, opts, is, physicalID, isIndex, histIDs, tablePartitionStats, h.getTableByPhysicalID, h.loadTablePartitionStats)
 }
 
 func (h *Handle) loadTablePartitionStats(tableInfo *model.TableInfo, partitionDef *model.PartitionDefinition) (*statistics.Table, error) {
@@ -408,269 +362,8 @@ func (h *Handle) loadTablePartitionStats(tableInfo *model.TableInfo, partitionDe
 func (h *Handle) mergePartitionStats2GlobalStats(sc sessionctx.Context,
 	opts map[ast.AnalyzeOptionType]uint64, is infoschema.InfoSchema, globalTableInfo *model.TableInfo,
 	isIndex int, histIDs []int64,
-	allPartitionStats map[int64]*statistics.Table) (globalStats *GlobalStats, err error) {
-	partitionNum := len(globalTableInfo.Partition.Definitions)
-
-	// initialized the globalStats
-	globalStats = new(GlobalStats)
-	if len(histIDs) == 0 {
-		for _, col := range globalTableInfo.Columns {
-			// The virtual generated column stats can not be merged to the global stats.
-			if col.IsGenerated() && !col.GeneratedStored {
-				continue
-			}
-			histIDs = append(histIDs, col.ID)
-		}
-	}
-	globalStats.Num = len(histIDs)
-	globalStats.Count = 0
-	globalStats.Hg = make([]*statistics.Histogram, globalStats.Num)
-	globalStats.Cms = make([]*statistics.CMSketch, globalStats.Num)
-	globalStats.TopN = make([]*statistics.TopN, globalStats.Num)
-	globalStats.Fms = make([]*statistics.FMSketch, globalStats.Num)
-
-	// The first dimension of slice is means the number of column or index stats in the globalStats.
-	// The second dimension of slice is means the number of partition tables.
-	// Because all topN and histograms need to be collected before they can be merged.
-	// So we should store all of the partition-level stats first, and merge them together.
-	allHg := make([][]*statistics.Histogram, globalStats.Num)
-	allCms := make([][]*statistics.CMSketch, globalStats.Num)
-	allTopN := make([][]*statistics.TopN, globalStats.Num)
-	allFms := make([][]*statistics.FMSketch, globalStats.Num)
-	for i := 0; i < globalStats.Num; i++ {
-		allHg[i] = make([]*statistics.Histogram, 0, partitionNum)
-		allCms[i] = make([]*statistics.CMSketch, 0, partitionNum)
-		allTopN[i] = make([]*statistics.TopN, 0, partitionNum)
-		allFms[i] = make([]*statistics.FMSketch, 0, partitionNum)
-	}
-
-	skipMissingPartitionStats := sc.GetSessionVars().SkipMissingPartitionStats
-	if sc.GetSessionVars().InRestrictedSQL {
-		// For AutoAnalyze and HandleDDLEvent(ActionDropTablePartition), we need to use @@global.tidb_skip_missing_partition_stats
-		val, err1 := sc.GetSessionVars().GlobalVarsAccessor.GetGlobalSysVar(variable.TiDBAnalyzeSkipColumnTypes)
-		if err1 != nil {
-			logutil.BgLogger().Error("loading tidb_skip_missing_partition_stats failed", zap.Error(err1))
-			err = err1
-			return
-		}
-		skipMissingPartitionStats = variable.TiDBOptOn(val)
-	}
-	for _, def := range globalTableInfo.Partition.Definitions {
-		partitionID := def.ID
-		partitionTable, ok := h.getTableByPhysicalID(is, partitionID)
-		if !ok {
-			err = errors.Errorf("unknown physical ID %d in stats meta table, maybe it has been dropped", partitionID)
-			return
-		}
-		tableInfo := partitionTable.Meta()
-		var partitionStats *statistics.Table
-		if allPartitionStats != nil {
-			partitionStats, ok = allPartitionStats[partitionID]
-		}
-		// If pre-load partition stats isn't provided, then we load partition stats directly and set it into allPartitionStats
-		if allPartitionStats == nil || partitionStats == nil || !ok {
-			var err1 error
-			partitionStats, err1 = h.loadTablePartitionStats(tableInfo, &def)
-			if err1 != nil {
-				if skipMissingPartitionStats && types.ErrPartitionStatsMissing.Equal(err) {
-					globalStats.MissingPartitionStats = append(globalStats.MissingPartitionStats, fmt.Sprintf("partition `%s`", def.Name.L))
-					continue
-				}
-				err = err1
-				return
-			}
-			if allPartitionStats == nil {
-				allPartitionStats = make(map[int64]*statistics.Table)
-			}
-			allPartitionStats[partitionID] = partitionStats
-		}
-		for i := 0; i < globalStats.Num; i++ {
-			hg, cms, topN, fms, analyzed := partitionStats.GetStatsInfo(histIDs[i], isIndex == 1)
-			skipPartition := false
-			if !analyzed {
-				var missingPart string
-				if isIndex == 0 {
-					missingPart = fmt.Sprintf("partition `%s` column `%s`", def.Name.L, tableInfo.FindColumnNameByID(histIDs[i]))
-				} else {
-					missingPart = fmt.Sprintf("partition `%s` index `%s`", def.Name.L, tableInfo.FindIndexNameByID(histIDs[i]))
-				}
-				if !skipMissingPartitionStats {
-					err = types.ErrPartitionStatsMissing.GenWithStackByArgs(fmt.Sprintf("table `%s` %s", tableInfo.Name.L, missingPart))
-					return
-				}
-				globalStats.MissingPartitionStats = append(globalStats.MissingPartitionStats, missingPart)
-				skipPartition = true
-			}
-			// partition stats is not empty but column stats(hist, topn) is missing
-			if partitionStats.RealtimeCount > 0 && (hg == nil || hg.TotalRowCount() <= 0) && (topN == nil || topN.TotalCount() <= 0) {
-				var missingPart string
-				if isIndex == 0 {
-					missingPart = fmt.Sprintf("partition `%s` column `%s`", def.Name.L, tableInfo.FindColumnNameByID(histIDs[i]))
-				} else {
-					missingPart = fmt.Sprintf("partition `%s` index `%s`", def.Name.L, tableInfo.FindIndexNameByID(histIDs[i]))
-				}
-				if !skipMissingPartitionStats {
-					err = types.ErrPartitionColumnStatsMissing.GenWithStackByArgs(fmt.Sprintf("table `%s` %s", tableInfo.Name.L, missingPart))
-					return
-				}
-				globalStats.MissingPartitionStats = append(globalStats.MissingPartitionStats, missingPart+" hist and topn")
-				skipPartition = true
-			}
-			if i == 0 {
-				// In a partition, we will only update globalStats.Count once
-				globalStats.Count += partitionStats.RealtimeCount
-				globalStats.ModifyCount += partitionStats.ModifyCount
-			}
-			if !skipPartition {
-				allHg[i] = append(allHg[i], hg)
-				allCms[i] = append(allCms[i], cms)
-				allTopN[i] = append(allTopN[i], topN)
-				allFms[i] = append(allFms[i], fms)
-			}
-		}
-	}
-
-	// After collect all of the statistics from the partition-level stats,
-	// we should merge them together.
-	for i := 0; i < globalStats.Num; i++ {
-		if len(allHg[i]) == 0 {
-			// If all partitions have no stats, we skip merging global stats because it may not handle the case `len(allHg[i]) == 0`
-			// correctly. It can avoid unexpected behaviors such as nil pointer panic.
-			continue
-		}
-		// Merge CMSketch
-		globalStats.Cms[i] = allCms[i][0].Copy()
-		for j := 1; j < len(allCms[i]); j++ {
-			err = globalStats.Cms[i].MergeCMSketch(allCms[i][j])
-			if err != nil {
-				return
-			}
-		}
-
-		// Merge topN. We need to merge TopN before merging the histogram.
-		// Because after merging TopN, some numbers will be left.
-		// These remaining topN numbers will be used as a separate bucket for later histogram merging.
-		var popedTopN []statistics.TopNMeta
-		wrapper := statistics.NewStatsWrapper(allHg[i], allTopN[i])
-		globalStats.TopN[i], popedTopN, allHg[i], err = mergeGlobalStatsTopN(h.gpool, sc, wrapper, sc.GetSessionVars().StmtCtx.TimeZone, sc.GetSessionVars().AnalyzeVersion, uint32(opts[ast.AnalyzeOptNumTopN]), isIndex == 1)
-		if err != nil {
-			return
-		}
-
-		// Merge histogram
-		globalStats.Hg[i], err = statistics.MergePartitionHist2GlobalHist(sc.GetSessionVars().StmtCtx, allHg[i], popedTopN, int64(opts[ast.AnalyzeOptNumBuckets]), isIndex == 1)
-		if err != nil {
-			return
-		}
-
-		// NOTICE: after merging bucket NDVs have the trend to be underestimated, so for safe we don't use them.
-		for j := range globalStats.Hg[i].Buckets {
-			globalStats.Hg[i].Buckets[j].NDV = 0
-		}
-
-		// Update NDV of global-level stats
-		globalStats.Fms[i] = allFms[i][0].Copy()
-		for j := 1; j < len(allFms[i]); j++ {
-			globalStats.Fms[i].MergeFMSketch(allFms[i][j])
-		}
-
-		// update the NDV
-		globalStatsNDV := globalStats.Fms[i].NDV()
-		if globalStatsNDV > globalStats.Count {
-			globalStatsNDV = globalStats.Count
-		}
-		globalStats.Hg[i].NDV = globalStatsNDV
-	}
-	return
-}
-
-func mergeGlobalStatsTopN(gp *gp.Pool, sc sessionctx.Context, wrapper *statistics.StatsWrapper,
-	timeZone *time.Location, version int, n uint32, isIndex bool) (*statistics.TopN,
-	[]statistics.TopNMeta, []*statistics.Histogram, error) {
-	mergeConcurrency := sc.GetSessionVars().AnalyzePartitionMergeConcurrency
-	killed := &sc.GetSessionVars().Killed
-	// use original method if concurrency equals 1 or for version1
-	if mergeConcurrency < 2 {
-		return statistics.MergePartTopN2GlobalTopN(timeZone, version, wrapper.AllTopN, n, wrapper.AllHg, isIndex, killed)
-	}
-	batchSize := len(wrapper.AllTopN) / mergeConcurrency
-	if batchSize < 1 {
-		batchSize = 1
-	} else if batchSize > MaxPartitionMergeBatchSize {
-		batchSize = MaxPartitionMergeBatchSize
-	}
-	return MergeGlobalStatsTopNByConcurrency(gp, mergeConcurrency, batchSize, wrapper, timeZone, version, n, isIndex, killed)
-}
-
-// MergeGlobalStatsTopNByConcurrency merge partition topN by concurrency
-// To merge global stats topn by concurrency, we will separate the partition topn in concurrency part and deal it with different worker.
-// mergeConcurrency is used to control the total concurrency of the running worker, and mergeBatchSize is sued to control
-// the partition size for each worker to solve it
-func MergeGlobalStatsTopNByConcurrency(gp *gp.Pool, mergeConcurrency, mergeBatchSize int, wrapper *statistics.StatsWrapper,
-	timeZone *time.Location, version int, n uint32, isIndex bool, killed *uint32) (*statistics.TopN,
-	[]statistics.TopNMeta, []*statistics.Histogram, error) {
-	if len(wrapper.AllTopN) < mergeConcurrency {
-		mergeConcurrency = len(wrapper.AllTopN)
-	}
-	tasks := make([]*statistics.TopnStatsMergeTask, 0)
-	for start := 0; start < len(wrapper.AllTopN); {
-		end := start + mergeBatchSize
-		if end > len(wrapper.AllTopN) {
-			end = len(wrapper.AllTopN)
-		}
-		task := statistics.NewTopnStatsMergeTask(start, end)
-		tasks = append(tasks, task)
-		start = end
-	}
-	var wg sync.WaitGroup
-	taskNum := len(tasks)
-	taskCh := make(chan *statistics.TopnStatsMergeTask, taskNum)
-	respCh := make(chan *statistics.TopnStatsMergeResponse, taskNum)
-	for i := 0; i < mergeConcurrency; i++ {
-		worker := statistics.NewTopnStatsMergeWorker(taskCh, respCh, wrapper, killed)
-		wg.Add(1)
-		gp.Go(func() {
-			defer wg.Done()
-			worker.Run(timeZone, isIndex, n, version)
-		})
-	}
-	for _, task := range tasks {
-		taskCh <- task
-	}
-	close(taskCh)
-	wg.Wait()
-	close(respCh)
-	resps := make([]*statistics.TopnStatsMergeResponse, 0)
-
-	// handle Error
-	hasErr := false
-	errMsg := make([]string, 0)
-	for resp := range respCh {
-		if resp.Err != nil {
-			hasErr = true
-			errMsg = append(errMsg, resp.Err.Error())
-		}
-		resps = append(resps, resp)
-	}
-	if hasErr {
-		return nil, nil, nil, errors.New(strings.Join(errMsg, ","))
-	}
-
-	// fetch the response from each worker and merge them into global topn stats
-	sorted := make([]statistics.TopNMeta, 0, mergeConcurrency)
-	leftTopn := make([]statistics.TopNMeta, 0)
-	for _, resp := range resps {
-		if resp.TopN != nil {
-			sorted = append(sorted, resp.TopN.TopN...)
-		}
-		leftTopn = append(leftTopn, resp.PopedTopn...)
-	}
-
-	globalTopN, popedTopn := statistics.GetMergedTopNFromSortedSlice(sorted, n)
-
-	result := append(leftTopn, popedTopn...)
-	statistics.SortTopnMeta(result)
-	return globalTopN, result, wrapper.AllHg, nil
+	allPartitionStats map[int64]*statistics.Table) (globalStats *globalstats.GlobalStats, err error) {
+	return h.globalstatushandler.MergePartitionStats2GlobalStats(sc, opts, is, globalTableInfo, isIndex, histIDs, allPartitionStats, h.getTableByPhysicalID, h.loadTablePartitionStats)
 }
 
 func (h *Handle) getTableByPhysicalID(is infoschema.InfoSchema, physicalID int64) (table.Table, bool) {
@@ -2033,6 +1726,6 @@ func (h *Handle) SetStatsCacheCapacity(c int64) {
 
 // Close stops the background
 func (h *Handle) Close() {
-	h.gpool.Close()
+	h.globalstatushandler.Close()
 	h.statsCache.Load().Close()
 }
